Moves the children of the ColumnHeader previously passed as props inside of the ColumHeader component.
